works or maybe does at least not raising exceptions
