+++ smallworld/analyses/__init__.py
@@ -2,12 +2,14 @@ from .analysis import *  # noqa: F401, F403
 from .analysis import __all__ as __analysis__
 from .colorizer import Colorizer
 from .colorizer_summary import ColorizerSummary
+from .colorizer_def_use import ColorizerDefUse
 from .field_detection import FieldDetectionAnalysis, ForcedFieldDetectionAnalysis
 from .forced_exec import ForcedExecution
 
 __all__ = __analysis__ + [
     "Colorizer",
     "ColorizerSummary",
+    "ColorizerDefUse",
     "FieldDetectionAnalysis",
     "ForcedFieldDetectionAnalysis",
     "ForcedExecution",
